Buyer Controller

+++ ApplicationService/DTOs/BuyerDTO.cs
@@ -14,7 +14,7 @@ namespace ApplicationService.DTOs
 
         public byte Age { get; set; }
 
-        public double Money { get; set; }
+        public decimal Money { get; set; }
 
         public int PhoneNumber { get; set; }
